back/requirements.txt
alembic==1.14.1
bcrypt==4.2.1
cryptography==44.0.0
DateTime==5.5
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Pillow==10.4.0
pipreqs==0.5.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.37
sqlalchemy-schemadisplay==2.0
SQLAlchemy-Utils==0.41.2
stripe==11.5.0
discord_webhook
